+++ src/Attribute/HasTabindex.php
@@ -36,10 +36,10 @@ trait HasTabindex
     /**
      * Sets the HTML `tabindex` attribute for the element.
      *
-     * Creates a new instance with the specified tab order value, supporting both explicit and nullable assignment
-     * according to the HTML specification for global attributes.
+     * Creates a new instance with the specified tab order, supporting both explicit and nullable assignment according
+     * to the HTML specification for global attributes.
      *
-     * @param int|string|null $value Tab order value to set for the element. Can be `null` to unset the attribute.
+     * @param int|string|null $value Tab order to set for the element. Can be `null` to unset the attribute.
      *
      * @throws InvalidArgumentException if the value is not a valid integer or string representation of an
      * `value >= -1`.