+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ChangeLog
 
-## 0.2.1 Under development
+## 0.3.0 Under development
 
 - Bug #6: Update testsuite name to match project name (@terabytesoftw)
 - Enh #7: Add methods to set and remove single HTML attributes (@terabytesoftw)
@@ -11,7 +11,8 @@
 - Bug #12: Update `removeDataAttribute()` method documentation to include example with enum key (@terabytesoftw)
 - Bug #13: Correct documentation formatting for key features across multiple attribute and providers external classes (@terabytesoftw)
 - Enh #14: Enhance `addAttribute()` and `removeAttribute()` methods to support enum keys and improve handling of attributes (@terabytesoftw)
-- Bug 315: Update PHPStan annotations in `DataProvider` and `HasDataTest` classes for improved type clarity (@terabytesoftw)
+- Bug #15: Update PHPStan annotations in `DataProvider` and `HasDataTest` classes for improved type clarity (@terabytesoftw)
+- Enh #16: Enhance support for `Stringable` interface types across `HasClass`, `HasData`, `HasStyle`, `HasTitle` and update related tests (@terabytesoftw)
 
 ## 0.2.0 December 18, 2025